Move IndividualCustomerPage to its parent folder for consistency
