+++ src/domain/individualCustomer/individualCustomerPage.module.scss
